friender/src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.2
decorator==5.1.1
Django==4.2.13
django-bootstrap-v5==1.0.11
django-celery-results==2.5.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==24.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
Faker==25.1.0
flower==2.0.1
humanize==4.9.0
ipython==8.24.0
jedi==0.19.1
kombu==5.3.7
Markdown==3.6
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.36
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
